+++ packages/openapi-ts/src/plugins/@hey-api/client-angular/bundle/client.ts
@@ -66,7 +66,6 @@ export const createClient = (config: Config = {}): Client => {
       ...options,
       headers: mergeHeaders(_config.headers, options.headers),
       httpClient: options.httpClient ?? _config.httpClient,
-      method: 'GET',
       serializedBody: options.body as any,
     };
 
@@ -93,7 +92,7 @@ export const createClient = (config: Config = {}): Client => {
     const url = buildUrl(opts as any);
 
     const req = new HttpRequest<unknown>(
-      opts.method,
+      opts.method ?? 'GET',
       url,
       opts.serializedBody || null,
       {
@@ -134,26 +133,28 @@ export const createClient = (config: Config = {}): Client => {
       }
     }
 
-    let response;
-    const result = {
+    const result: {
+      request: HttpRequest<unknown>;
+      response: any;
+    } = {
       request: req,
-      response,
+      response: null,
     };
 
     try {
-      response = await firstValueFrom(
+      result.response = (await firstValueFrom(
         opts
           .httpClient!.request(req)
           .pipe(filter((event) => event.type === HttpEventType.Response)),
-      );
+      )) as HttpResponse<unknown>;
 
       for (const fn of interceptors.response._fns) {
         if (fn) {
-          response = await fn(response, req, opts as any);
+          result.response = await fn(result.response, req, opts as any);
         }
       }
 
-      let bodyResponse: any = response.body;
+      let bodyResponse = result.response.body;
 
       if (opts.responseValidator) {
         await opts.responseValidator(bodyResponse);
@@ -168,7 +169,7 @@ export const createClient = (config: Config = {}): Client => {
         : { data: bodyResponse, ...result };
     } catch (error) {
       if (error instanceof HttpErrorResponse) {
-        response = error;
+        result.response = error;
       }
 
       let finalError = error instanceof HttpErrorResponse ? error.error : error;
@@ -177,7 +178,7 @@ export const createClient = (config: Config = {}): Client => {
         if (fn) {
           finalError = (await fn(
             finalError,
-            response as HttpResponse<unknown>,
+            result.response as any,
             req,
             opts as any,
           )) as string;